Make dedicated options panel
+ Rename the old HuntOptions to HuntFilters
